crt int value updated

+++ NIBMCafeteria/View Controllers/Food/CartCell.swift
@@ -8,11 +8,11 @@
 import UIKit
 
 protocol addItemsAmtDelegate {
-    func addAmtItems(amount:Double?,index:IndexPath?)
+    func addAmtItems(amount:Int?,index:IndexPath?)
 }
 
 protocol minItemsAmtDelegate {
-    func minAmtItems(amount:Double?,index:IndexPath?)
+    func minAmtItems(amount:Int?,index:IndexPath?)
 
 }
